requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylama==8.4.1
python-dotenv==1.0.1
requests==2.31.0
snowballstemmer==2.2.0
urllib3==2.2.1
Werkzeug==3.0.2
